simbricks: add PCI adapter
Adds a gem5 adapter for the SimBricks PCI interface. To gem5 this
is just another implementaiton of a PCI device. There is a bit of
trickery in the implementation to make sure that timing protocol memory
accesses can be handled fully asynchronously, as they are forwarded to
external simulators and completed asynchronously.

Change-Id: If87a301879bfd6665c24850d308fbeca96a39724
Co-authored-by: Jialin Li <lijl@comp.nus.edu.sg>